only the user who created the question can archive it

+++ app/Http/Controllers/Question/ArchiveController.php
@@ -4,12 +4,16 @@ namespace App\Http\Controllers\Question;
 
 use App\Http\Controllers\Controller;
 use App\Models\Question;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ArchiveController extends Controller
 {
+    use AuthorizesRequests;
     public function __invoke(Question $question)
     {
+        $this->authorize('archive', $question);
         $question->delete();
+
         return response()->noContent();
     }
 }
